moves one more trace function to general utils area
